+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-- Fix: Icons not lining up at their base. 
+## [5.0.0](https://github.com/rizen/pixi-glyphs/compare/v4.3.0...v5.0.0) (2025-10-24)
+
+
+### Bug Fixes
+
+* Icons not lining up at their base. ([86eda58](https://github.com/rizen/pixi-glyphs/commit/86eda588d40378512e014a6d3a577e0d619e4201))
 
 ## [4.3.0](https://github.com/rizen/pixi-glyphs/compare/v4.2.1...v4.3.0) (2025-10-23)